requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# LLM providers
openai>=1.3.0
anthropic>=0.18.0  # Optional: for Anthropic Claude support
groq>=0.4.0  # Optional: for Groq fast inference support

# Data and storage
duckdb>=0.9.0
pandas>=2.1.0

# UI
streamlit>=1.28.0

# Templating and reporting
jinja2>=3.1.2
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
presidio-analyzer>=2.2.0  # Optional PII detection

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
black>=23.10.0
httpx>=0.25.0  # For testing FastAPI
requests>=2.31.0  # For demo script

# Typing
types-PyYAML>=6.0.0
